# performance changes